refactor(#964): split spawn_single_instance + tests to sibling files
CI surfaced `file_size_invariant` violation: `src/mcp/handlers/instance.rs`
exceeded the 750-LOC ceiling (970 LOC) after the swap fix + inline
tests in commits 2454919 + e3b6545.

Split per the existing `channel_p0a_tests.rs` / `p0b_tests.rs` /
`comms_p0c_tests.rs` precedent:

- `src/mcp/handlers/instance_spawn.rs` (new, 230 LOC) — relocated
  `spawn_single_instance` + `spawn_single_instance_impl` (the swap
  fix) + `rollback_fleet_entry_on_failure` (the rollback helper).
  `handle_create_instance` at `instance.rs:118` now calls
  `super::instance_spawn::spawn_single_instance(...)`.

- `src/mcp/handlers/instance_964_tests.rs` (new, 183 LOC) —
  relocated `tests_964` module (T1 + T2 + T2b). Imports
  `super::instance_spawn::spawn_single_instance_impl` instead of
  `super::spawn_single_instance_impl`.

- `src/mcp/handlers/mod.rs` — registered both new modules. The test
  module is gated by `#[cfg(test)]`; the spawn module is plain
  `mod instance_spawn;` (private to `handlers`).

Post-split sizes:
  instance.rs            563 LOC (under 750 ✓)
  instance_spawn.rs      230 LOC
  instance_964_tests.rs  183 LOC

Behavior identical to e3b6545. All 4 #964 tests pass; `file_size_invariant`
test now passes locally.

## Local gates
- cargo fmt --check ✓
- cargo clippy --bin agend-terminal --all-targets --features tray
  -- -D warnings ✓
- cargo test --bin agend-terminal --features tray: 2511 pass / 0 fail ✓
- file_size_invariant: passes ✓

Refs #964

Agend-Agent: fixup-dev-2
Agend-Task: t-20260520011500-964-impl
Agend-Branch: fix/964-mcp-create-instance-ordering
Agend-Issued-At: 2026-05-19T17:19:51+00:00
