+++ src/mcp/handlers/instance.rs
@@ -1,6 +1,4 @@
-use crate::agent_ops::{
-    list_agents, merge_metadata, save_metadata, save_metadata_batch, validate_branch,
-};
+use crate::agent_ops::{list_agents, merge_metadata, save_metadata, save_metadata_batch};
 use crate::identity::Sender;
 use serde_json::{json, Value};
 use std::path::Path;
@@ -115,7 +113,7 @@ pub(super) fn handle_create_instance(home: &Path, args: &Value, instance_name: &
             Err(e) => json!({"error": format!("API unavailable: {e}")}),
         }
     } else {
-        spawn_single_instance(home, instance_name, args)
+        super::instance_spawn::spawn_single_instance(home, instance_name, args)
     }
 }
 
@@ -558,413 +556,6 @@ pub(super) fn resolve_team_layout(
     (layout, target)
 }
 
-fn spawn_single_instance(home: &Path, instance_name: &str, args: &Value) -> Value {
-    spawn_single_instance_impl(home, instance_name, args, &crate::api::call)
-}
-
-/// Inner impl of [`spawn_single_instance`] parameterized on the SPAWN RPC.
-///
-/// The `spawn_fn` indirection exists for the T1/T2 regression tests: they
-/// pass a mock that mimics `handle_spawn`'s side effects (the SPAWN-time
-/// `register_topic` → `update_instance_field` chain) without standing up
-/// a real daemon. Production passes [`crate::api::call`].
-fn spawn_single_instance_impl(
-    home: &Path,
-    instance_name: &str,
-    args: &Value,
-    spawn_fn: &dyn Fn(&Path, &Value) -> anyhow::Result<Value>,
-) -> Value {
-    let raw_name = match args["name"].as_str() {
-        Some(n) => n,
-        None => return json!({"error": "missing 'name'"}),
-    };
-    if let Err(e) = crate::agent::validate_name(raw_name) {
-        return json!({"error": e});
-    }
-    let name_owned = {
-        // M4: AtomicU64 prevents 65536 wrap-around collision
-        use std::sync::atomic::{AtomicU64, Ordering};
-        static DEDUP_SEQ: AtomicU64 = AtomicU64::new(0);
-
-        let existing: std::collections::HashSet<String> =
-            crate::fleet::FleetConfig::load(&crate::fleet::fleet_yaml_path(home))
-                .map(|c| c.instance_names().into_iter().collect())
-                .unwrap_or_default();
-        if existing.contains(raw_name) {
-            let seq = DEDUP_SEQ.fetch_add(1, Ordering::Relaxed);
-            let deduped = format!("{raw_name}-{seq:04x}");
-            tracing::info!(original = raw_name, deduped = %deduped, "name conflict, auto-deduped");
-            deduped
-        } else {
-            raw_name.to_string()
-        }
-    };
-    let name: &str = &name_owned;
-    let command = args["backend"]
-        .as_str()
-        .or_else(|| args["command"].as_str())
-        .unwrap_or("claude");
-    let mut cmd_args = args
-        .get("args")
-        .and_then(|v| v.as_str())
-        .filter(|s| !s.is_empty())
-        .map(String::from)
-        .unwrap_or_default();
-    if let Some(model) = args
-        .get("model")
-        .and_then(|v| v.as_str())
-        .filter(|m| !m.is_empty())
-    {
-        let model_val = crate::backend::Backend::from_command(command)
-            .map(|b| b.format_model_arg(model))
-            .unwrap_or_else(|| model.to_string());
-        if !cmd_args.is_empty() {
-            cmd_args.push(' ');
-        }
-        cmd_args.push_str(&format!("--model {model_val}"));
-    }
-    if let Some(dir) = args.get("working_directory").and_then(|v| v.as_str()) {
-        if dir.contains("..") {
-            return json!({"error": "working_directory must not contain '..'"});
-        }
-        if !dir.starts_with('/') {
-            return json!({"error": "working_directory must be an absolute path"});
-        }
-    }
-    let mut work_dir = args
-        .get("working_directory")
-        .and_then(|v| v.as_str())
-        .map(String::from)
-        .unwrap_or_else(|| {
-            crate::paths::workspace_dir(home)
-                .join(name)
-                .display()
-                .to_string()
-        });
-
-    if let Some(branch) = args.get("branch").and_then(|v| v.as_str()) {
-        if !validate_branch(branch) {
-            return json!({"error": format!("invalid branch name '{branch}'")});
-        }
-        let wd = std::path::PathBuf::from(&work_dir);
-        // Sprint 57 Wave 4 (#546 Item 4): worktree creation now takes
-        // `home` so the canonical external layout
-        // `$AGEND_HOME/worktrees/<agent>/<branch>/` resolves correctly.
-        if let Some(info) = crate::worktree::create(home, &wd, name, Some(branch)) {
-            work_dir = info.path.display().to_string();
-        }
-    }
-
-    std::fs::create_dir_all(&work_dir).ok();
-
-    let task = args.get("task").and_then(|v| v.as_str()).map(String::from);
-    let role = args.get("role").and_then(|v| v.as_str()).map(String::from);
-    let backend_str = args
-        .get("backend")
-        .and_then(|v| v.as_str())
-        .map(String::from);
-    // #900: forward operator-supplied `env` through the SPAWN RPC AND
-    // record it on the fleet.yaml entry. The runtime payload lets the
-    // daemon's handle_spawn apply it directly (no second fleet.yaml
-    // read); the persisted entry covers replace_instance / restart
-    // flows that re-resolve from disk later. Non-string values are
-    // filtered out at the daemon side via `parse_env_object`.
-    let env_value: Option<Value> = args.get("env").filter(|v| v.is_object()).cloned();
-    let env_for_entry: Option<std::collections::HashMap<String, String>> =
-        env_value.as_ref().and_then(|v| {
-            v.as_object().map(|obj| {
-                obj.iter()
-                    .filter_map(|(k, val)| val.as_str().map(|s| (k.clone(), s.to_string())))
-                    .collect()
-            })
-        });
-    let (layout, target_pane_owned) =
-        resolve_team_layout(home, name, args.get("layout"), args.get("target_pane"));
-    let target_pane = target_pane_owned.as_deref();
-
-    // #964 fix: build fleet.yaml entry + ADD FIRST, BEFORE the SPAWN RPC.
-    // Pre-fix ordering was SPAWN → add, which meant `handle_spawn`'s
-    // `register_topic` → `update_instance_field("topic_id")` ran while the
-    // fleet.yaml entry didn't yet exist — #962 Path 2 silent no-op fired
-    // and `topic_id` was never persisted. The 27-day-latent bug only
-    // surfaced when an instance was deleted before any daemon-restart-time
-    // self-heal (`channel/telegram/bootstrap.rs:126-150`) could backfill
-    // the missing field.
-    let entry = crate::fleet::InstanceYamlEntry {
-        backend: backend_str
-            .or_else(|| {
-                crate::backend::Backend::from_command(command).map(|b| b.name().to_string())
-            })
-            .or_else(|| Some(command.to_string())),
-        working_directory: Some(work_dir.clone()),
-        role: role.clone(),
-        instructions: None,
-        // Sprint 54 P1-B Bug 2 fix: gradient deployment per
-        // general's constraint — daemon auto-write leaves
-        // source_repo None; operator hand-edits fleet.yaml to
-        // opt agents in. Backward-compat preserved via
-        // dispatch_auto_bind_lease's working_directory fallback.
-        source_repo: None,
-        // Sprint 55 P0-B EC4: same gradient — daemon leaves
-        // `repo` override None; operator opts in for non-GitHub
-        // remote OR fork-vs-upstream disambiguation.
-        repo: None,
-        github_login: None,
-        args: None,
-        model: None,
-        env: env_for_entry,
-        ready_pattern: None,
-        command: None,
-        worktree: None,
-    };
-    if let Err(e) = crate::fleet::add_instance_to_yaml(home, name, &entry) {
-        return json!({"error": format!("failed to register instance in fleet.yaml: {e}")});
-    }
-
-    let mut spawn_params = json!({
-        "name": name, "backend": command, "args": &cmd_args,
-        "working_directory": work_dir,
-        "layout": layout, "spawner": instance_name,
-        "target_pane": target_pane,
-        "role": role,
-    });
-    if let Some(env) = env_value.as_ref() {
-        spawn_params["env"] = env.clone();
-    }
-    match spawn_fn(
-        home,
-        &json!({"method": crate::api::method::SPAWN, "params": spawn_params}),
-    ) {
-        Ok(resp) if resp["ok"].as_bool() == Some(true) => {
-            let topic_id = resp["result"]["topic_id"].as_i64();
-            if let Some(task_text) = task {
-                let h = home.to_path_buf();
-                let n = name.to_string();
-                // fire-and-forget: single-agent task injection (M5 §10.5).
-                std::thread::Builder::new()
-                    .name("task_inject".into())
-                    .spawn(move || {
-                        std::thread::sleep(std::time::Duration::from_secs(3));
-                        let _ = crate::api::call(
-                            &h,
-                            &json!({"method": crate::api::method::INJECT, "params": {"name": n, "data": task_text}}),
-                        );
-                    })
-                    .ok();
-            }
-            let mut result = json!({"name": name, "backend": command});
-            if let Some(tid) = topic_id {
-                result["topic_id"] = json!(tid);
-            }
-            result
-        }
-        Ok(resp) => {
-            // #964 rollback: SPAWN reported failure — undo the fleet.yaml
-            // entry we just added so create_instance is all-or-nothing.
-            // dev-2 cross-audit Pushback 1: surface rollback failure via
-            // `tracing::error!` (NOT `let _ = ...` — that would repeat the
-            // #962 antipattern). Operator gets an audit trail on the rare
-            // double-failure case.
-            if let Err(remove_err) = crate::fleet::remove_instance_from_yaml(home, name) {
-                tracing::error!(
-                    name = %name,
-                    error = %remove_err,
-                    "create_instance: SPAWN failed AND rollback failed — \
-                     fleet.yaml may have stale entry; operator may need manual cleanup"
-                );
-            }
-            json!({"error": resp["error"].as_str().unwrap_or("spawn failed")})
-        }
-        Err(e) => {
-            if let Err(remove_err) = crate::fleet::remove_instance_from_yaml(home, name) {
-                tracing::error!(
-                    name = %name,
-                    error = %remove_err,
-                    "create_instance: API unavailable AND rollback failed — \
-                     fleet.yaml may have stale entry; operator may need manual cleanup"
-                );
-            }
-            json!({"error": format!("API unavailable: {e}")})
-        }
-    }
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used)]
-mod tests_964 {
-    //! #964 regression tests — caller-side integration coverage for the
-    //! SPAWN-then-add ordering bug that was latent since PR #417 (May 4)
-    //! and only surfaced after #962 made the silent no-op observable.
-    //!
-    //! T1 + T2 inject a mock `spawn_fn` into `spawn_single_instance_impl`
-    //! that mimics `handle_spawn`'s side effects (`register_topic` writes
-    //! `topic_id` back via `update_instance_field`) WITHOUT standing up a
-    //! real daemon. This isolates the ordering invariant: the fleet.yaml
-    //! entry MUST exist when SPAWN runs, otherwise the mock's
-    //! `update_instance_field` call hits #962 Path 2 (silent no-op) and
-    //! the assertion fails — which is exactly what HEAD-pre-fix did.
-    //!
-    //! Class lesson (per dev-2 cross-audit): the existing test
-    //! `topic_id_persists_to_fleet_yaml_via_update_instance_field`
-    //! (api/handlers/instance.rs:712, added with PR #417) is helper-
-    //! level and never exercised the caller-path ordering. These tests
-    //! pin the caller-path contract — what PR #417 should have shipped.
-    use super::*;
-
-    fn tmp_home(slug: &str) -> std::path::PathBuf {
-        use std::sync::atomic::{AtomicU64, Ordering};
-        static SEQ: AtomicU64 = AtomicU64::new(0);
-        let id = SEQ.fetch_add(1, Ordering::Relaxed);
-        let dir =
-            std::env::temp_dir().join(format!("agend-964-{}-{}-{}", slug, std::process::id(), id));
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).unwrap();
-        std::fs::write(crate::fleet::fleet_yaml_path(&dir), "instances: {}\n").unwrap();
-        dir
-    }
-
-    /// T1 (load-bearing): MCP create_instance must persist `topic_id` to
-    /// fleet.yaml on the happy path. The mock `spawn_fn` mimics
-    /// `handle_spawn`'s `register_topic` chain: it calls
-    /// `update_instance_field("topic_id", ...)` then returns the topic_id
-    /// in the SPAWN response. After the swap fix, the fleet.yaml entry
-    /// exists by the time the mock fires, so the helper write succeeds and
-    /// the post-state has `topic_id = Some(5198)`.
-    ///
-    /// Pre-#964 (SPAWN-then-add ordering): mock would fire BEFORE the
-    /// caller added the entry → helper hits #962 Path 2 silent no-op →
-    /// `topic_id` stays `None` → assertion fails. This test is the
-    /// regression anchor.
-    #[test]
-    fn t1_create_instance_persists_topic_id_to_fleet_yaml() {
-        let home = tmp_home("t1");
-        let target_topic_id = 5198_i32;
-
-        let spawn_fn = |h: &Path, req: &Value| -> anyhow::Result<Value> {
-            let name = req["params"]["name"]
-                .as_str()
-                .ok_or_else(|| anyhow::anyhow!("missing name"))?;
-            // Mimic handle_spawn → register_topic → update_instance_field.
-            // With the swap fix, fleet.yaml entry exists here, so the
-            // helper persists topic_id. Pre-fix, the entry didn't exist,
-            // so the helper silently no-op'd.
-            crate::fleet::update_instance_field(
-                h,
-                name,
-                "topic_id",
-                serde_yaml_ng::Value::Number(serde_yaml_ng::Number::from(target_topic_id)),
-            )?;
-            Ok(json!({
-                "ok": true,
-                "result": {"topic_id": target_topic_id}
-            }))
-        };
-
-        let result = spawn_single_instance_impl(
-            &home,
-            "test-spawner",
-            &json!({"name": "t1-instance", "backend": "claude"}),
-            &spawn_fn,
-        );
-
-        assert_eq!(
-            result["topic_id"].as_i64(),
-            Some(target_topic_id as i64),
-            "MCP response must carry topic_id; got {result}"
-        );
-
-        let cfg = crate::fleet::FleetConfig::load(&crate::fleet::fleet_yaml_path(&home))
-            .expect("reload fleet.yaml");
-        let persisted = cfg.instances.get("t1-instance").and_then(|i| i.topic_id);
-        assert_eq!(
-            persisted,
-            Some(target_topic_id),
-            "#964 regression: topic_id must be persisted to fleet.yaml \
-             after MCP create_instance returns. Got {persisted:?} for \
-             entry={:?}",
-            cfg.instances.get("t1-instance")
-        );
-
-        let _ = std::fs::remove_dir_all(&home);
-    }
-
-    /// T2 (rollback): when SPAWN reports failure, the fleet.yaml entry
-    /// that was added pre-SPAWN MUST be rolled back so create_instance is
-    /// all-or-nothing. Without rollback, a failed create would leave a
-    /// stale entry that confuses subsequent list/replace flows.
-    #[test]
-    fn t2_spawn_failure_rolls_back_fleet_yaml_entry() {
-        let home = tmp_home("t2");
-
-        let spawn_fn = |_h: &Path, _req: &Value| -> anyhow::Result<Value> {
-            Ok(json!({"ok": false, "error": "backend rejected spawn"}))
-        };
-
-        let result = spawn_single_instance_impl(
-            &home,
-            "test-spawner",
-            &json!({"name": "t2-instance", "backend": "claude"}),
-            &spawn_fn,
-        );
-
-        assert!(
-            result.get("error").is_some(),
-            "SPAWN failure must surface as error; got {result}"
-        );
-        assert!(
-            result["error"]
-                .as_str()
-                .unwrap_or("")
-                .contains("backend rejected spawn"),
-            "error must propagate inner SPAWN error; got {result}"
-        );
-
-        let cfg = crate::fleet::FleetConfig::load(&crate::fleet::fleet_yaml_path(&home))
-            .expect("reload fleet.yaml");
-        assert!(
-            !cfg.instance_names().iter().any(|n| n == "t2-instance"),
-            "#964 rollback: fleet.yaml must NOT contain t2-instance \
-             after SPAWN failure. Got instances={:?}",
-            cfg.instance_names()
-        );
-
-        let _ = std::fs::remove_dir_all(&home);
-    }
-
-    /// T2b (rollback on API-unavailable): same contract as T2, but the
-    /// spawn_fn returns `Err` (API-unavailable path) instead of an
-    /// `Ok(json!{"ok": false})` response.
-    #[test]
-    fn t2b_api_unavailable_rolls_back_fleet_yaml_entry() {
-        let home = tmp_home("t2b");
-
-        let spawn_fn =
-            |_h: &Path, _req: &Value| -> anyhow::Result<Value> { Err(anyhow::anyhow!("ipc dead")) };
-
-        let result = spawn_single_instance_impl(
-            &home,
-            "test-spawner",
-            &json!({"name": "t2b-instance", "backend": "claude"}),
-            &spawn_fn,
-        );
-
-        assert!(
-            result["error"]
-                .as_str()
-                .unwrap_or("")
-                .contains("API unavailable"),
-            "Err path must surface as 'API unavailable'; got {result}"
-        );
-
-        let cfg = crate::fleet::FleetConfig::load(&crate::fleet::fleet_yaml_path(&home))
-            .expect("reload fleet.yaml");
-        assert!(
-            !cfg.instance_names().iter().any(|n| n == "t2b-instance"),
-            "#964 rollback on API-unavailable: fleet.yaml must NOT \
-             contain t2b-instance; got instances={:?}",
-            cfg.instance_names()
-        );
-
-        let _ = std::fs::remove_dir_all(&home);
-    }
-}
+// #964: spawn_single_instance + spawn_single_instance_impl live in
+// sibling file `instance_spawn.rs` to keep this file under the 750-LOC
+// file_size_invariant.
